Various update for v1.1.1

+++ ModdingAssetUpdater/Program.cs
@@ -44,53 +44,60 @@ foreach (var (name, path) in ZipFiles)
     await zipArchive.ExtractToDirectoryAsync(outputPath);
 }
 
-// 3. Copy the ripped assets
-FrozenSet<string> CopyingExtensions = ((string[])[".prefab", ".mat", ".png"]).ToFrozenSet(StringComparer.OrdinalIgnoreCase);
-Stack<string> rippedPaths = new([""]);
-
-Console.WriteLine("Copying ripped assets...");
-while (rippedPaths.Count > 0)
+if (!Directory.Exists(AssetRipperFolder))
 {
-    var subPath = rippedPaths.Pop();
-    var fullPath = Path.Combine(AssetRipperFolder, subPath);
-
-    foreach (var subFolder in Directory.EnumerateDirectories(fullPath))
-    {
-        var name = Path.GetFileName(subFolder);
-        rippedPaths.Push(Path.Combine(subPath, name));
-    }
-
-    var outputPath = Path.Combine(OutputFolder, subPath);
-    Directory.CreateDirectory(outputPath);
-    foreach (var file in Directory.EnumerateFiles(fullPath))
-    {
-        var ext = Path.GetExtension(file);
-        if (!CopyingExtensions.Contains(ext)) { continue; }
-
-        var name = Path.GetFileName(file);
-        Console.WriteLine($"Copying {name} to {outputPath}");
-        File.Copy(file, Path.Combine(outputPath, name));
-    }
+    Console.WriteLine($"AssetRipperFolder '{AssetRipperFolder}' does not exist, skipping copying ripped assets.");
 }
-
-// 4. Copy Placeholder Shaders to map later
-Console.WriteLine("Copying placeholder shaders...");
-
-var rippedShaderFolder = Path.Combine(AssetRipperFolder, "../Shader");
-var dstPlaceholderShader = Path.Combine(OutputFolder, "ShaderPlaceholders");
-Directory.CreateDirectory(dstPlaceholderShader);
-
-foreach (var path in Directory.EnumerateFiles(rippedShaderFolder))
+else
 {
-    var fileName = Path.GetFileName(path);
-    var dst = Path.Combine(dstPlaceholderShader, fileName);
-    File.Copy(path, dst);
+    // 3. Copy the ripped assets
+    FrozenSet<string> CopyingExtensions = ((string[])[".prefab", ".mat", ".png"]).ToFrozenSet(StringComparer.OrdinalIgnoreCase);
+    Stack<string> rippedPaths = new([""]);
 
-    if (!fileName.EndsWith(".shader")) { continue; }
+    Console.WriteLine("Copying ripped assets...");
+    while (rippedPaths.Count > 0)
+    {
+        var subPath = rippedPaths.Pop();
+        var fullPath = Path.Combine(AssetRipperFolder, subPath);
 
-    var lines = await File.ReadAllLinesAsync(dst);
-    lines[0] = lines[0].Replace("\" {", "2\" {");
-    await File.WriteAllLinesAsync(dst, lines);
+        foreach (var subFolder in Directory.EnumerateDirectories(fullPath))
+        {
+            var name = Path.GetFileName(subFolder);
+            rippedPaths.Push(Path.Combine(subPath, name));
+        }
+
+        var outputPath = Path.Combine(OutputFolder, subPath);
+        Directory.CreateDirectory(outputPath);
+        foreach (var file in Directory.EnumerateFiles(fullPath))
+        {
+            var ext = Path.GetExtension(file);
+            if (!CopyingExtensions.Contains(ext)) { continue; }
+
+            var name = Path.GetFileName(file);
+            Console.WriteLine($"Copying {name} to {outputPath}");
+            File.Copy(file, Path.Combine(outputPath, name));
+        }
+    }
+
+    // 4. Copy Placeholder Shaders to map later
+    Console.WriteLine("Copying placeholder shaders...");
+
+    var rippedShaderFolder = Path.Combine(AssetRipperFolder, "../Shader");
+    var dstPlaceholderShader = Path.Combine(OutputFolder, "ShaderPlaceholders");
+    Directory.CreateDirectory(dstPlaceholderShader);
+
+    foreach (var path in Directory.EnumerateFiles(rippedShaderFolder))
+    {
+        var fileName = Path.GetFileName(path);
+        var dst = Path.Combine(dstPlaceholderShader, fileName);
+        File.Copy(path, dst);
+
+        if (!fileName.EndsWith(".shader")) { continue; }
+
+        var lines = await File.ReadAllLinesAsync(dst);
+        lines[0] = lines[0].Replace("\" {", "2\" {");
+        await File.WriteAllLinesAsync(dst, lines);
+    }
 }
 
 // 5. Decompile game DLLs
